+++ index.ts
@@ -1,8 +1,26 @@
-import { httpServer } from "./src/http_server/index.js";
 import dotenv from 'dotenv'
-
+import * as fs from 'fs';
+import * as path from 'path';
+import * as http from 'http';
+import {startWs} from './src/http_server/server'
 dotenv.config();
-const HTTP_PORT = 8181;
+
+export const httpServer = http.createServer(function (req, res) {
+    const __dirname = path.resolve(path.dirname(''));
+    const file_path = __dirname + (req.url === '/' ? '/front/index.html' : '/front' + req.url);
+    fs.readFile(file_path, function (err, data) {
+        if (err) {
+            res.writeHead(404);
+            res.end(JSON.stringify(err));
+            return;
+        }
+        res.writeHead(200);
+        res.end(data);
+    });
+});
+
+const HTTP_PORT = 8000;
 
 console.log(`Start static http server on the ${HTTP_PORT} port!`);
 httpServer.listen(HTTP_PORT);
+startWs();
